Refactor and rename go tool
